+++ lib/generators/classy_enum/templates/enum.rb
@@ -1,5 +1,4 @@
 class <%= class_name %> < ClassyEnum::Base
-  enum_classes <%= values.map {|a| ":#{a}"}.join(", ") %>
 end
 <% values.each do |arg| %>
 class <%= "#{class_name}::#{arg.camelize}" %> < <%= class_name %>