+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # ClassyEnum Changelog
 
+## 3.0.0
+
+* Removing ClassyEnum::Base.enum_classes in favor of using enum
+  inheritance to setup classes
+* Removing ClassyEnum::Base.valid_options
+* Removing ClassyEnum::Base.find
+* Removing ClassyEnum::Base#name
+* Removing :suffix option from classy_enum_attr
+* Enforce use of namespacing for subclasses (Parent::Child < Parent)
+* Use require instead of autoload
+* Lots of code refactoring
+
 ## 2.3.0
 
 * Deprecating ClassyEnum::Base#name (use to_s.titleize instead). `name` is